add more C89 and portability clauses to wolfsentry_settings.h, including new feature gates WOLFSENTRY_HAVE_DESIGNATED_INITIALIZERS, WOLFSENTRY_HAVE_LONG_LONG, and SIZE_T_32/SIZE_T_64, and rearrange to get FreeRTOS build to succeed with C89;
add c89-test to Makefile.analyzers and add it to check deps;

add m32-c89-test and freertos-arm32-c89-build-test and add them to check-extra deps;

remove stale/unneeded gcov exclusions;

transform miscellaneous C++-style comments into C-style;

add !WOLFSENTRY_HAVE_DESIGNATED_INITIALIZERS and !WOLFSENTRY_HAVE_LONG_LONG code paths where needed;

fix several static_asserts() that were meant to be wolfsentry_static_asserts().

+++ wolfsentry/wolfsentry_util.h
@@ -57,8 +57,8 @@
 #define streq(vs,fs,vs_len) (((vs_len) == strlen(fs)) && (memcmp(vs,fs,vs_len) == 0))
 #define strcaseeq(vs,fs,vs_len) (((vs_len) == strlen(fs)) && (strncasecmp(vs,fs,vs_len) == 0))
 
-#define MAX_UINT_OF(x) (((1ULL << ((sizeof(x) * (unsigned long long)BITS_PER_BYTE) - 1ULL)) - 1ULL) | (1ULL << ((sizeof(x) * BITS_PER_BYTE) - 1ULL)))
-#define MAX_SINT_OF(x) (((1LL << ((sizeof(x) * (long long)BITS_PER_BYTE) - 2LL)) - 1LL) | (1LL << ((sizeof(x) * BITS_PER_BYTE) - 2LL)))
+#define MAX_UINT_OF(x) ((((uint64_t)1 << ((sizeof(x) * (uint64_t)BITS_PER_BYTE) - (uint64_t)1)) - (uint64_t)1) | ((uint64_t)1 << ((sizeof(x) * (uint64_t)BITS_PER_BYTE) - (uint64_t)1)))
+#define MAX_SINT_OF(x) ((int64_t)((((uint64_t)1 << ((sizeof(x) * (uint64_t)BITS_PER_BYTE) - (uint64_t)2)) - (uint64_t)1) | ((uint64_t)1 << ((sizeof(x) * (uint64_t)BITS_PER_BYTE) - (uint64_t)2))))
 
 #define WOLFSENTRY_SET_BITS(enumint, bits) ((enumint) |= (bits))
 #define WOLFSENTRY_CHECK_BITS(enumint, bits) (((enumint) & (bits)) == (bits))
